+++ coroutines/src/commonMain/kotlin/dev/inmo/micro_utils/coroutines/HandleSafely.kt
@@ -39,9 +39,11 @@ object ContextSafelyExceptionHandlerKey : CoroutineContext.Key<ContextSafelyExce
  */
 class ContextSafelyExceptionHandler(
     val handler: ExceptionHandler<Unit>
-) : CoroutineContext.Element {
+) : CoroutineContext.Element, CoroutineExceptionHandler {
     override val key: CoroutineContext.Key<*>
         get() = ContextSafelyExceptionHandlerKey
+
+    override fun handleException(context: CoroutineContext, exception: Throwable) = handler(exception)
 }
 
 /**
@@ -147,3 +149,10 @@ suspend inline fun <T> runCatchingSafelyWithoutExceptions(
 ): Result<T?> = runCatching {
     safelyWithoutExceptions(onException, block)
 }
+
+suspend inline fun CoroutineScope(
+    context: CoroutineContext,
+    defaultExceptionsHandler: ExceptionHandler<Unit>
+) = CoroutineScope(
+    context + ContextSafelyExceptionHandler(defaultExceptionsHandler)
+)